Correct the codes related to enum
By now the enum subclasses check statement in visitInnerClass() of Class
MockedClassModifier might not be suitable, which will cause test method
mockEnumWithValueSpecificMethods of MockedEnumsTest failure at JUnit
Test, the source of the problem is that inner class
EnumWithValueSpecificMethods.One,Two are not redefined as their access
value are 0x4000(ENUM) - tested on jdk 8/13/15.
The 'access == ENUM + STATIC' statement in isSyntheticOrEnumClass() of
Class CoverageModifier might be wrong because it didn't confirm to Bits
operation practice.